requirements.txt
Django==4.0.6
djangorestframework==3.13.1
ffmpeg-python==0.2.0
